# api-docs.json
{"openapi":"3.1.0","info":{"title":"The Todolist API Backend built @ Bbc","version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"Authorization":[]}],"paths":{"/tags":{"get":{"tags":["tag-controller"],"summary":"Get all tags / Get tags by name","operationId":"findTags","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Tags found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TagResponseDTO"}}}},"404":{"description":"Tags were not found"}}},"post":{"tags":["tag-controller"],"summary":"Create a tag","operationId":"insert","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TagRequestDTO"}}},"required":true},"responses":{"201":{"description":"tag inserted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TagRequestDTO"}}}},"400":{"description":"tag was not inserted"}}}},"/items":{"get":{"tags":["item-controller"],"summary":"Get all items / Get items by name / Get items by tag name / Get items by name and tag name","operationId":"findItems","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}},{"name":"tagName","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Items found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ItemResponseDTO"}}}},"404":{"description":"Items were not found"}}},"post":{"tags":["item-controller"],"summary":"Create an item","operationId":"insert_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemRequestDTO"}}},"required":true},"responses":{"201":{"description":"Item inserted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ItemRequestDTO"}}}},"400":{"description":"Bad request"}}}},"/auth/signup":{"post":{"tags":["auth-controller"],"summary":"Create a new user","operationId":"signUp","requestBody":{"description":"The user to register","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequestDTO"}}},"required":true},"responses":{"201":{"description":"User was created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponseDTO"}}}},"409":{"description":"User could not be created, username already in use"}},"security":[]}},"/auth/signin":{"post":{"tags":["auth-controller"],"summary":"Receive a token for BEARER authorization","operationId":"signIn","requestBody":{"description":"The user to login","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequestDTO"}}},"required":true},"responses":{"200":{"description":"Login successful","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtResponseDTO"}}}},"401":{"description":"Invalid credentials"}},"security":[]}},"/tags/{id}":{"get":{"tags":["tag-controller"],"summary":"Get a Tag","operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Tag found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TagResponseDTO"}}}},"404":{"description":"Tag was not found"}}},"delete":{"tags":["tag-controller"],"summary":"Delete a tag","operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"204":{"description":"Tag was deleted successfully"},"404":{"description":"Tag could not be deleted"}}},"patch":{"tags":["tag-controller"],"summary":"Update a tag","operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"description":"The tag to update","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TagResponseDTO"}}},"required":true},"responses":{"200":{"description":"Tag was updated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TagResponseDTO"}}}},"404":{"description":"Tag was not found"},"409":{"description":"There was a conflict while updating the tag"}}}},"/persons/{id}":{"get":{"tags":["person-controller"],"summary":"Get a person","operationId":"findById_1","parameters":[{"name":"id","in":"path","description":"Id of person to get","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Person found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PersonResponseDTO"}}}},"404":{"description":"Person was not found"}}},"delete":{"tags":["person-controller"],"summary":"Delete a person","operationId":"delete_1","parameters":[{"name":"id","in":"path","description":"Id of person to delete","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"204":{"description":"Person was deleted successfully"},"404":{"description":"Person could not be deleted"}}},"patch":{"tags":["person-controller"],"summary":"Update a person","operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"description":"The person to update","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonRequestDTO"}}},"required":true},"responses":{"200":{"description":"Person was updated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PersonResponseDTO"}}}},"404":{"description":"Person was not found"},"409":{"description":"There was a conflict while updating the person"}}}},"/items/{id}":{"get":{"tags":["item-controller"],"summary":"Get an Item","operationId":"findById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Item found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ItemResponseDTO"}}}},"404":{"description":"Item was not found"}}},"delete":{"tags":["item-controller"],"summary":"Delete an item","operationId":"delete_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"204":{"description":"Item was deleted successfully"},"404":{"description":"Item could not be deleted"}}},"patch":{"tags":["item-controller"],"summary":"Update an item","operationId":"update_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"description":"The item to update","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemResponseDTO"}}},"required":true},"responses":{"200":{"description":"Item was updated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ItemResponseDTO"}}}},"404":{"description":"Item was not found"},"409":{"description":"There was a conflict while updating the item"}}}},"/persons":{"get":{"tags":["person-controller"],"summary":"Get all persons.","operationId":"findAll","responses":{"200":{"description":"Persons found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PersonResponseDTO"}}}}},"404":{"description":"Persons not found"}}}}},"components":{"schemas":{"TagRequestDTO":{"type":"object","properties":{"name":{"type":"string","maxLength":256,"minLength":1}},"required":["name"]},"ItemRequestDTO":{"type":"object","properties":{"name":{"type":"string","maxLength":256,"minLength":1},"description":{"type":"string"},"personId":{"type":"integer","format":"int32"},"tagIds":{"type":"array","items":{"type":"integer","format":"int32"}},"done":{"type":"boolean"}},"required":["name","personId"]},"AuthRequestDTO":{"type":"object","properties":{"username":{"type":"string","maxLength":45,"minLength":0},"password":{"type":"string","maxLength":255,"minLength":8}},"required":["password","username"]},"AuthResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"username":{"type":"string"}}},"JwtResponseDTO":{"type":"object","properties":{"accessToken":{"type":"string"}}},"TagResponseDTO":{"type":"object","properties":{"name":{"type":"string","maxLength":256,"minLength":1},"id":{"type":"integer","format":"int32"},"itemIds":{"type":"array","items":{"type":"integer","format":"int32"}}},"required":["name"]},"PersonRequestDTO":{"type":"object","properties":{"username":{"type":"string","minLength":1}},"required":["username"]},"PersonResponseDTO":{"type":"object","properties":{"username":{"type":"string","minLength":1},"id":{"type":"integer","format":"int32"},"itemIds":{"type":"array","items":{"type":"integer","format":"int32"}}},"required":["username"]},"ItemResponseDTO":{"type":"object","properties":{"name":{"type":"string","maxLength":256,"minLength":1},"description":{"type":"string"},"personId":{"type":"integer","format":"int32"},"tagIds":{"type":"array","items":{"type":"integer","format":"int32"}},"done":{"type":"boolean"},"id":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"}},"required":["name","personId"]}},"securitySchemes":{"Authorization":{"type":"http","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}